Tests/LibWeb/Text/expected/HTML/window-open-tokenized-features.txt

aliases:
locationbar is BarProp: true visible: false
menubar is BarProp: true visible: false
personalbar is BarProp: true visible: false
scrollbars is BarProp: true visible: false
statusbar is BarProp: true visible: false
toolbar is BarProp: true visible: false
not popup:
locationbar is BarProp: true visible: true
menubar is BarProp: true visible: true
personalbar is BarProp: true visible: true
scrollbars is BarProp: true visible: true
statusbar is BarProp: true visible: true
toolbar is BarProp: true visible: true
